+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This changelog documents all notable changes of the RealMQ Platform.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Introduce new administrative user management endpoints `/admin/v1/realms/{realmId}/users`.
+
 ### Security
 - Upgraded node to version 16 ([#114])
 - Upgraded dependencies ([#116])